Move shared presentation contracts and text helpers
- Move mode, help-action, and prompt contracts into `presentation`.
- Import reusable terminal text operations directly from `ag-tui-text`.
- Update architecture and maintenance guidance for the new module ownership.
