+++ crates/agentty/src/presentation.rs
@@ -1,8 +1,5 @@
 //! Frontend-neutral interaction state shared by runtime input and UI output.
 
-#[path = "ui/state/app_mode.rs"]
 pub mod app_mode;
-#[path = "ui/state/help_action.rs"]
 pub mod help_action;
-#[path = "ui/state/prompt.rs"]
 pub mod prompt;
