+++ crates/agentty/src/presentation/app_mode.rs
