+++ crates/agentty/src/presentation/help_action.rs
