+++ crates/agentty/src/presentation/prompt.rs
