+++ crates/agentty/src/ui/AGENTS.md
@@ -9,9 +9,9 @@ When working within `crates/agentty/src/ui/`:
   (`session_list.rs`, `session_chat.rs`, etc.).
 - **Helper Functions**: Keep render functions focused. Move reusable or complex layout
   and text-processing helpers to the narrowest shared module that fits the behavior (for
-  example, `layout.rs`, `input_layout.rs`, `prompt_format.rs`, `question_format.rs`,
-  `session_format.rs`, or `text_util.rs`) and cover that helper with a unit test when
-  practical.
+  example, `layout.rs`, `input_layout.rs`, `prompt_format.rs`, `question_format.rs`, or
+  `session_format.rs`) and cover that helper with a unit test when practical. Put
+  reusable terminal text operations in `ag-tui-text`.
 - **Component Reuse**: Check the `component/` directory before building a new common
   widget. All components must implement the `Component` trait.
 - **Palette Usage**: Use semantic color tokens from `style.rs` (`palette::*`) for UI
@@ -61,11 +61,12 @@ When working within `crates/agentty/src/ui/`:
 - `render.rs` and `router.rs` own frame composition and page dispatch.
 - `page.rs` and `page/` own full-screen pages.
 - `component.rs` and `component/` own reusable widgets and overlays.
-- `state.rs` and `state/` own UI mode and prompt state.
+- `presentation.rs` and `presentation/` own frontend-neutral mode, help-action, and
+  prompt state shared by runtime input and UI output.
 - `layout.rs` owns pure area and panel geometry.
 - `input_layout.rs` owns chat-input wrapping, cursor geometry, dropdown sizing, and
   table-column geometry.
 - `prompt_format.rs`, `question_format.rs`, and `session_format.rs` own prompt,
   clarification-question, and session display formatting.
-- `style.rs`, `markdown.rs`, `text_util.rs`, `diff_util.rs`, and `activity_heatmap.rs`
-  own focused shared presentation helpers.
+- `style.rs`, `markdown.rs`, `diff_util.rs`, and `activity_heatmap.rs` own focused
+  shared presentation helpers; reusable text operations come from `ag-tui-text`.
