+++ crates/agentty/src/ui/component/session_output.rs
@@ -4,6 +4,7 @@ use std::fmt::Write as _;
 use std::hash::Hasher;
 use std::sync::Arc;
 
+use ag_tui_text::text_util;
 use ratatui::Frame;
 use ratatui::buffer::Buffer;
 use ratatui::layout::Rect;
@@ -20,7 +21,7 @@ use crate::ui::icon::{Icon, TACHYON_LOADER_WIDTH};
 use crate::ui::input_layout::{bottom_pinned_scroll_offset, panel_inner_width};
 use crate::ui::markdown::{self, render_markdown};
 use crate::ui::prompt_block::{self, USER_PROMPT_PREFIX, USER_PROMPT_RIGHT_GUTTER_WIDTH};
-use crate::ui::{Component, session_format, style, text_util};
+use crate::ui::{Component, session_format, style};
 
 const DRAFT_PREVIEW_HEADER: &str = "## Draft Session";
 const DRAFT_PREVIEW_EMPTY_NOTE: &str = "No draft messages staged yet. Use `Enter` to stage the \
