+++ crates/agentty/src/ui/page/inbox.rs
@@ -1,4 +1,5 @@
 use ag_forge::{RequestedReview, RequestedReviewAudience};
+use ag_tui_text::text_util::{inline_text, truncate_spans_with_ellipsis};
 use ratatui::Frame;
 use ratatui::layout::{Constraint, Rect};
 use ratatui::style::{Modifier, Style};
@@ -8,7 +9,6 @@ use ratatui::widgets::{Block, Borders, Cell, Paragraph, Row, Table, TableState};
 use crate::app::RequestedReviewState;
 use crate::presentation::help_action;
 use crate::ui::input_layout::first_table_column_width;
-use crate::ui::text_util::{inline_text, truncate_spans_with_ellipsis};
 use crate::ui::{Page, layout, style};
 
 /// Horizontal spacing between requested-review table columns.
