+++ crates/agentty/src/ui/prompt_block.rs
@@ -1,9 +1,10 @@
 //! Shared transcript prompt-block layout constants.
 
+use ag_tui_text::text_util;
 use ratatui::style::{Modifier, Style};
 use ratatui::text::{Line, Span};
 
-use crate::ui::{style, text_util};
+use crate::ui::style;
 
 /// Visible prefix for the first row of a user prompt in transcript output.
 pub(crate) const USER_PROMPT_PREFIX: &str = " › ";
