--- crates/agentty/src/ui/state.rs
@@ -1,3 +1,0 @@
-//! Compatibility re-exports for shared presentation state.
-
-pub use crate::presentation::{app_mode, help_action, prompt};
