--- crates/agentty/src/ui/text_util.rs
@@ -1,6 +1,0 @@
-pub(crate) use ag_tui_text::text_util::{borrowed_paint_line, borrowed_paint_lines};
-pub use ag_tui_text::text_util::{
-    format_duration_compact, format_token_count, inline_text, split_trailing_line_block,
-    truncate_spans_with_ellipsis, truncate_with_ellipsis, wrap_lines, wrap_lines_to_rows,
-    wrap_styled_line, wrapped_line_count,
-};
